+++ app/controllers/questions_controller.rb
@@ -1,10 +1,12 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
 
   def index
     @questions = Question.all
   end
 
   def show
+    @answer = question.answers.new
   end
 
   def new
@@ -15,9 +17,9 @@ class QuestionsController < ApplicationController
 
   def create
     @question = Question.new(question_params)
-
+    @question.author = current_user
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -32,8 +34,8 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    question.destroy if current_user.author_of?(question)
+    redirect_to questions_path, notice: 'The Question was successfully deleted.'
   end
 
   private
